+++ angular/lib/platform/browser_static.dart
@@ -1,5 +1,0 @@
-/// Import 'package:angular/angular.dart' instead.
-@deprecated
-library angular.platform.browser;
-
-export 'package:angular/angular.dart';